Codigo incluido en proyecto de code::blocks
