+++ pedidos/articulos.h
@@ -1,0 +1,5 @@
+#include <stdio.h>
+#include <string.h>
+
+void listadoArticulos();
+
