+++ pedidos/fecha.h
@@ -1,0 +1,4 @@
+#include <stdio.h>
+#include <time.h>
+
+void fechaYHora();
